backend/requirements.txt
Flask==3.1.1
flask_sqlalchemy==3.1.1
PyJWT==2.10.1
pytest==8.4.1
SQLAlchemy==2.0.41
psycopg2==2.9.10
flask_bcrypt==1.0.1
flask_cors==6.0.1
flask_debugtoolbar==0.16.0
flask_migrate==4.1.0
gunicorn==23.0.0
boto3==1.35.65
pillow==11.0.0
python-dotenv==1.1.1